Introduce Directory as a separate type
The key motivation here is the ability to represent the tags within a
directory separate from a file. This allows us to represent directories
that are not image directories, i.e. the EXIF directory, the related GPS
directory, and other sub directories. These are defined by a 'pointer'
value in another tag, an offset into the file. They can then be read
from that immediate location.